모든 코드 및 폴더 정리
