--- src/App.css
@@ -1,28 +1,0 @@
-.App {
-    text-align: center;
-    min-height: 100%;
-}
-.App-logo {
-height: 40vmin;
-pointer-events: none;
-}
-.spinner{
-position: "fixed";
-top: 50%;
-left: 50%;
-}
-
-
-@media (prefers-reduced-motion: no-preference) {
-.App-logo {
-    animation: App-logo-spin infinite 20s linear;
-}
-}
-
-.body-wrapper {
-height: auto;
-min-height: 100vh;
-padding-bottom:70px;
-background : linear-gradient(to bottom, #ced1f8 30%, #ebf2ff, #ebf4ff);
-}
-
