+++ src/App.module.css
@@ -1,62 +1,25 @@
-#index {
-    z-index: 1;
-}
-.sticky {
-    position: sticky;
-    top: 0;
-    z-index: 10;
-    background-color: #212529;
-}
-
-#subCommunity{
-    background-color: white;
-}
-#box{
-    border: 1px solid grey;
-    background-color: ghostwhite;
-    background-attachment: fixed;
-}
-.top100 {
-    font-size: 3em;
-    font-weight: bold;
-    color: ghostwhite;
-    text-align: left;
-    text-shadow: 4px 4px 1px #818be0, -1px -1px 1px #838de7;
-}
-
-.btn {
-    position: relative;
-    top: 7px;
-
-    width: 50px;
-    height: 50px;
-
-    border: 2px solid rgb(125, 138, 255);
-    color: rgb(125, 138, 255);
+.App {
     text-align: center;
-    border-radius: 50%;
-    background-color: rgba(0, 0, 0, 0);
+    min-height: 100%;
 }
-.btn:hover {
-    background-color: rgb(202, 199, 255);
+.App-logo {
+    height: 40vmin;
+    pointer-events: none;
 }
-
-.space {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    
-    border-radius: 20px;
-    background-color: rgba(255, 252, 248, 0.8);
-    box-shadow: 1px 1px 5px rgb(126, 126, 126), 0px 1px 5px rgb(104, 104, 104);
-    z-index: 0;
+.spinner{
+    position: "fixed";
+    top: 50%;
+    left: 50%;
 }
-
-#wrapper {
-    position: relative;
-    left: 12px;
-    display:flex;
-    justify-content:center;
-    z-index: 1;
+.wrapper {
+    height: auto;
+    min-height: 100vh;
+    padding-bottom:70px;
+    background : linear-gradient(to bottom, #ced1f8 30%, #ebf2ff, #ebf4ff);
 }
 
+@media (prefers-reduced-motion: no-preference) {
+    .App-logo {
+        animation: App-logo-spin infinite 20s linear;
+    }
+}
