+++ src/essential/serviceWorkerRegistration.js
