+++ src/fixedComponent/Developer.js
@@ -1,5 +1,4 @@
 import React from 'react';
-
 export default function Developer() {
     return(
         <>
