+++ src/fixedComponent/Fixed.css
@@ -5,3 +5,9 @@
     padding: 10px;
     text-align: center;
 }
+.sticky {
+    position: sticky;
+    top: 0;
+    z-index: 10;
+    background-color: #212529;
+}
