+++ src/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 
-import * as serviceWorkerRegistration from './root/serviceWorkerRegistration';
-import reportWebVitals from './root/reportWebVitals';
+import * as serviceWorkerRegistration from './essential/serviceWorkerRegistration';
+import reportWebVitals from './essential/reportWebVitals';
 
 import './index.css';
 import App from './App';
