+++ src/music/Body.module.css
@@ -1,0 +1,52 @@
+#index {
+    z-index: 1;
+}
+#subCommunity{
+    background-color: white;
+}
+#box{
+    border: 1px solid grey;
+    background-color: ghostwhite;
+    background-attachment: fixed;
+}
+#wrapper {
+    position: relative;
+    left: 12px;
+    display:flex;
+    justify-content:center;
+    z-index: 1;
+}
+
+.top100 {
+    font-size: 3em;
+    font-weight: bold;
+    color: ghostwhite;
+    text-align: left;
+    text-shadow: 4px 4px 1px #818be0, -1px -1px 1px #838de7;
+}
+.btn {
+    position: relative;
+    top: 7px;
+
+    width: 50px;
+    height: 50px;
+
+    border: 2px solid rgb(125, 138, 255);
+    color: rgb(125, 138, 255);
+    text-align: center;
+    border-radius: 50%;
+    background-color: rgba(0, 0, 0, 0);
+}
+.btn:hover {
+    background-color: rgb(202, 199, 255);
+}
+.space {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    
+    border-radius: 20px;
+    background-color: rgba(255, 252, 248, 0.8);
+    box-shadow: 1px 1px 5px rgb(126, 126, 126), 0px 1px 5px rgb(104, 104, 104);
+    z-index: 0;
+}
