--- src/music/Chart.css
@@ -1,8 +1,0 @@
-.click-enter {
-    opacity : 0
-}
-  
-.click-enter-active {
-    opacity : 1;
-    transition : all 500ms;
-}
