+++ src/music/features/Slider.js
